+++ lib/components/Icon/svg/NotificationSvg.tsx
@@ -1,0 +1,17 @@
+/**
+ * DO NOT MODIFY THIS FILE MANUALLY
+ * This file is auto-generated via `yarn run generate:icons`
+ */
+/* tslint:disable:max-line-length */
+import * as React from 'react';
+
+const NotificationSvg = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg viewBox="0 0 24 24" {...props}>
+    <path
+      d="M14.5,6 L13,6 L13,8 L8.8,8 L8.8,8 C8.3581722,8 8,8.3581722 8,8.8 L8,15.2 C8,15.6418278 8.3581722,16 8.8,16 L15.2,16 C15.6418278,16 16,15.6418278 16,15.2 L16,11 L18,11 L18,9.5 L18,16 C18,17.1045695 17.1045695,18 16,18 L8,18 C6.8954305,18 6,17.1045695 6,16 L6,8 L6,8 C6,6.8954305 6.8954305,6 8,6 L8,6 L14.5,6 Z M16.5,9 C15.6715729,9 15,8.32842712 15,7.5 C15,6.67157288 15.6715729,6 16.5,6 C17.3284271,6 18,6.67157288 18,7.5 C18,8.32842712 17.3284271,9 16.5,9 Z"
+      id="Combined-Shape"
+    />
+  </svg>
+);
+
+export default NotificationSvg;