+++ lib/components/Icon/svg/SheetSvg.tsx
@@ -1,0 +1,17 @@
+/**
+ * DO NOT MODIFY THIS FILE MANUALLY
+ * This file is auto-generated via `yarn run generate:icons`
+ */
+/* tslint:disable:max-line-length */
+import * as React from 'react';
+
+const SheetSvg = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg viewBox="0 0 24 24" {...props}>
+    <path
+      d="M8,7.5 C7.72385763,7.5 7.5,7.72385763 7.5,8 L7.5,16 C7.5,16.2761424 7.72385763,16.5 8,16.5 L16,16.5 C16.2761424,16.5 16.5,16.2761424 16.5,16 L16.5,8 C16.5,7.72385763 16.2761424,7.5 16,7.5 L8,7.5 Z M8,6 L16,6 C17.1045695,6 18,6.8954305 18,8 L18,16 C18,17.1045695 17.1045695,18 16,18 L8,18 C6.8954305,18 6,17.1045695 6,16 L6,8 C6,6.8954305 6.8954305,6 8,6 Z M7,9.5 L17,9.5 L17,11 L7,11 L7,9.5 Z M12.8000002,6 L12.8000002,18 L11.3000002,18 L11.3000002,6 L12.8000002,6 Z M7,12.8999996 L17,12.8999996 L17,14.3999996 L7,14.3999996 L7,12.8999996 Z"
+      id="i--sheet"
+    />
+  </svg>
+);
+
+export default SheetSvg;