+++ lib/components/Icon/svg/index.ts
@@ -1,0 +1,200 @@
+/**
+ * DO NOT MODIFY THIS FILE MANUALLY
+ * This file is auto-generated via `yarn run generate:icons`
+ */
+
+import AppIconPlaceholderSvg from './AppIconPlaceholderSvg';
+import AppsSvg from './AppsSvg';
+import ArrowBackSvg from './ArrowBackSvg';
+import ArrowDownLeftSvg from './ArrowDownLeftSvg';
+import ArrowForwardSvg from './ArrowForwardSvg';
+import ArrowLeftSvg from './ArrowLeftSvg';
+import ArrowRightSvg from './ArrowRightSvg';
+import ArrowUpRightPlusSvg from './ArrowUpRightPlusSvg';
+import ArrowUpRightSvg from './ArrowUpRightSvg';
+import AtSvg from './AtSvg';
+import BellOffSvg from './BellOffSvg';
+import BellSvg from './BellSvg';
+import BurgerSvg from './BurgerSvg';
+import ChatSvg from './ChatSvg';
+import CheckmarkSvg from './CheckmarkSvg';
+import CheckroundSvg from './CheckroundSvg';
+import CogLineSvg from './CogLineSvg';
+import CogSvg from './CogSvg';
+import CrossSvg from './CrossSvg';
+import DetachSvg from './DetachSvg';
+import DocSvg from './DocSvg';
+import DownloadSvg from './DownloadSvg';
+import ExtensionSvg from './ExtensionSvg';
+import FocusSvg from './FocusSvg';
+import HeartfilledSvg from './HeartfilledSvg';
+import HeartstrokeSvg from './HeartstrokeSvg';
+import HintSvg from './HintSvg';
+import InfoSvg from './InfoSvg';
+import JobsSvg from './JobsSvg';
+import JoinSvg from './JoinSvg';
+import LeaveSvg from './LeaveSvg';
+import LifeboatSvg from './LifeboatSvg';
+import LinkChainSvg from './LinkChainSvg';
+import LoadingSvg from './LoadingSvg';
+import LockSvg from './LockSvg';
+import MarkReadSvg from './MarkReadSvg';
+import NotificationSvg from './NotificationSvg';
+import NudgeSvg from './NudgeSvg';
+import PencilSvg from './PencilSvg';
+import PinSvg from './PinSvg';
+import PlusSvg from './PlusSvg';
+import QuoteSvg from './QuoteSvg';
+import ReattachSvg from './ReattachSvg';
+import RecentArrowSvg from './RecentArrowSvg';
+import RecentSvg from './RecentSvg';
+import ReturnSvg from './ReturnSvg';
+import SearchSvg from './SearchSvg';
+import SendPaperplaneSvg from './SendPaperplaneSvg';
+import SendSvg from './SendSvg';
+import SheetSvg from './SheetSvg';
+import ShowSvg from './ShowSvg';
+import SlideSvg from './SlideSvg';
+import StarSvg from './StarSvg';
+import SupportSvg from './SupportSvg';
+import TextSvg from './TextSvg';
+import TimeSvg from './TimeSvg';
+import TrashSvg from './TrashSvg';
+import UnhappySvg from './UnhappySvg';
+import UpdateSvg from './UpdateSvg';
+import UserAddSvg from './UserAddSvg';
+import UserSvg from './UserSvg';
+
+export enum SvgSymbol {
+  APP_ICON_PLACEHOLDER = 'app-icon-placeholder',
+  APPS = 'apps',
+  ARROW_BACK = 'arrow-back',
+  ARROW_DOWN_LEFT = 'arrow-down-left',
+  ARROW_FORWARD = 'arrow-forward',
+  ARROW_LEFT = 'arrow-left',
+  ARROW_RIGHT = 'arrow-right',
+  ARROW_UP_RIGHT_PLUS = 'arrow-up-right-plus',
+  ARROW_UP_RIGHT = 'arrow-up-right',
+  AT = 'at',
+  BELL_OFF = 'bell-off',
+  BELL = 'bell',
+  BURGER = 'burger',
+  CHAT = 'chat',
+  CHECKMARK = 'checkmark',
+  CHECKROUND = 'checkround',
+  COG_LINE = 'cog-line',
+  COG = 'cog',
+  CROSS = 'cross',
+  DETACH = 'detach',
+  DOC = 'doc',
+  DOWNLOAD = 'download',
+  EXTENSION = 'extension',
+  FOCUS = 'focus',
+  HEARTFILLED = 'heartfilled',
+  HEARTSTROKE = 'heartstroke',
+  HINT = 'hint',
+  INFO = 'info',
+  JOBS = 'jobs',
+  JOIN = 'join',
+  LEAVE = 'leave',
+  LIFEBOAT = 'lifeboat',
+  LINK_CHAIN = 'link-chain',
+  LOADING = 'loading',
+  LOCK = 'lock',
+  MARK_READ = 'mark-read',
+  NOTIFICATION = 'notification',
+  NUDGE = 'nudge',
+  PENCIL = 'pencil',
+  PIN = 'pin',
+  PLUS = 'plus',
+  QUOTE = 'quote',
+  REATTACH = 'reattach',
+  RECENT_ARROW = 'recent-arrow',
+  RECENT = 'recent',
+  RETURN = 'return',
+  SEARCH = 'search',
+  SEND_PAPERPLANE = 'send-paperplane',
+  SEND = 'send',
+  SHEET = 'sheet',
+  SHOW = 'show',
+  SLIDE = 'slide',
+  STAR = 'star',
+  SUPPORT = 'support',
+  TEXT = 'text',
+  TIME = 'time',
+  TRASH = 'trash',
+  UNHAPPY = 'unhappy',
+  UPDATE = 'update',
+  USER_ADD = 'user-add',
+  USER = 'user',
+}
+
+type SvgComponents = {
+  [key in SvgSymbol]: React.SFC<React.SVGProps<SVGSVGElement>>;
+};
+
+const components: SvgComponents = {
+  [SvgSymbol.APP_ICON_PLACEHOLDER]: AppIconPlaceholderSvg,
+  [SvgSymbol.APPS]: AppsSvg,
+  [SvgSymbol.ARROW_BACK]: ArrowBackSvg,
+  [SvgSymbol.ARROW_DOWN_LEFT]: ArrowDownLeftSvg,
+  [SvgSymbol.ARROW_FORWARD]: ArrowForwardSvg,
+  [SvgSymbol.ARROW_LEFT]: ArrowLeftSvg,
+  [SvgSymbol.ARROW_RIGHT]: ArrowRightSvg,
+  [SvgSymbol.ARROW_UP_RIGHT_PLUS]: ArrowUpRightPlusSvg,
+  [SvgSymbol.ARROW_UP_RIGHT]: ArrowUpRightSvg,
+  [SvgSymbol.AT]: AtSvg,
+  [SvgSymbol.BELL_OFF]: BellOffSvg,
+  [SvgSymbol.BELL]: BellSvg,
+  [SvgSymbol.BURGER]: BurgerSvg,
+  [SvgSymbol.CHAT]: ChatSvg,
+  [SvgSymbol.CHECKMARK]: CheckmarkSvg,
+  [SvgSymbol.CHECKROUND]: CheckroundSvg,
+  [SvgSymbol.COG_LINE]: CogLineSvg,
+  [SvgSymbol.COG]: CogSvg,
+  [SvgSymbol.CROSS]: CrossSvg,
+  [SvgSymbol.DETACH]: DetachSvg,
+  [SvgSymbol.DOC]: DocSvg,
+  [SvgSymbol.DOWNLOAD]: DownloadSvg,
+  [SvgSymbol.EXTENSION]: ExtensionSvg,
+  [SvgSymbol.FOCUS]: FocusSvg,
+  [SvgSymbol.HEARTFILLED]: HeartfilledSvg,
+  [SvgSymbol.HEARTSTROKE]: HeartstrokeSvg,
+  [SvgSymbol.HINT]: HintSvg,
+  [SvgSymbol.INFO]: InfoSvg,
+  [SvgSymbol.JOBS]: JobsSvg,
+  [SvgSymbol.JOIN]: JoinSvg,
+  [SvgSymbol.LEAVE]: LeaveSvg,
+  [SvgSymbol.LIFEBOAT]: LifeboatSvg,
+  [SvgSymbol.LINK_CHAIN]: LinkChainSvg,
+  [SvgSymbol.LOADING]: LoadingSvg,
+  [SvgSymbol.LOCK]: LockSvg,
+  [SvgSymbol.MARK_READ]: MarkReadSvg,
+  [SvgSymbol.NOTIFICATION]: NotificationSvg,
+  [SvgSymbol.NUDGE]: NudgeSvg,
+  [SvgSymbol.PENCIL]: PencilSvg,
+  [SvgSymbol.PIN]: PinSvg,
+  [SvgSymbol.PLUS]: PlusSvg,
+  [SvgSymbol.QUOTE]: QuoteSvg,
+  [SvgSymbol.REATTACH]: ReattachSvg,
+  [SvgSymbol.RECENT_ARROW]: RecentArrowSvg,
+  [SvgSymbol.RECENT]: RecentSvg,
+  [SvgSymbol.RETURN]: ReturnSvg,
+  [SvgSymbol.SEARCH]: SearchSvg,
+  [SvgSymbol.SEND_PAPERPLANE]: SendPaperplaneSvg,
+  [SvgSymbol.SEND]: SendSvg,
+  [SvgSymbol.SHEET]: SheetSvg,
+  [SvgSymbol.SHOW]: ShowSvg,
+  [SvgSymbol.SLIDE]: SlideSvg,
+  [SvgSymbol.STAR]: StarSvg,
+  [SvgSymbol.SUPPORT]: SupportSvg,
+  [SvgSymbol.TEXT]: TextSvg,
+  [SvgSymbol.TIME]: TimeSvg,
+  [SvgSymbol.TRASH]: TrashSvg,
+  [SvgSymbol.UNHAPPY]: UnhappySvg,
+  [SvgSymbol.UPDATE]: UpdateSvg,
+  [SvgSymbol.USER_ADD]: UserAddSvg,
+  [SvgSymbol.USER]: UserSvg,
+};
+
+export default components;